reworked-project-structure
