--- Model/README.md
@@ -1,2 +1,0 @@
-#### Model Instructions
-Model is pre-trained and packaged as callable using joblib
