+++ pyproject.toml
@@ -11,5 +11,12 @@ dependencies = [
     "uvicorn>=0.35.0",
     "websockets>=15.0.1",
     "torch>=2.3.0",
-    "spacy>=3.8.0"
+    "spacy>=3.8.0",
+    "pandas>=2.0.0",
+    "numpy>=1.24.0",
+    "scikit-learn>=1.3.0",
+    "nltk>=3.8.0",
+    "pip>=25.2", 
+    "tqdm>=4.65.0",
+    "ipykernel>=6.25.0",
 ]
